refactor(api): Add subpackages in dto package
