+++ backend/FleetForge/src/main/java/com/team18/FleetForge/dto/driver/DriverPasswordRequestDTO.java
@@ -1,4 +1,4 @@
-package com.team18.FleetForge.dto;
+package com.team18.FleetForge.dto.driver;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,6 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class DriverPasswordResponseDTO {
-    long id;
-    private String email;
+public class DriverPasswordRequestDTO {
+    private String password;
 }
