+++ backend/FleetForge/src/main/java/com/team18/FleetForge/dto/driver/DriverPasswordResponseDTO.java
@@ -1,4 +1,4 @@
-package com.team18.FleetForge.dto;
+package com.team18.FleetForge.dto.driver;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class DriverPasswordRequestDTO {
-    private String password;
+public class DriverPasswordResponseDTO {
+    long id;
+    private String email;
 }
