+++ backend/FleetForge/src/main/java/com/team18/FleetForge/dto/ride/AdminRideHistoryItemDTO.java
@@ -1,5 +1,6 @@
-package com.team18.FleetForge.dto;
+package com.team18.FleetForge.dto.ride;
 
+import com.team18.FleetForge.dto.UserSummaryDTO;
 import lombok.Builder;
 import lombok.Data;
