+++ backend/FleetForge/src/main/java/com/team18/FleetForge/dto/ride/RideEstimateRequestDTO.java
@@ -1,4 +1,4 @@
-package com.team18.FleetForge.dto;
+package com.team18.FleetForge.dto.ride;
 
 import com.team18.FleetForge.model.GeoPoint;
 import lombok.Data;
